requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
dataclasses-json==0.6.7
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.128.8
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langsmith==0.4.37
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.3.0
ollama==0.6.1
orjson==3.11.5
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
SQLAlchemy==2.0.48
starlette==0.49.3
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
yarl==1.22.0
zstandard==0.25.0
